Struct tea::io::Reader

pub struct Reader<R: BufRead> {
    // some fields omitted
}

Wraps an underlying std::io::BufRead so that bytes read get decrypted on the way through.

Example:

use std::fs::File;
use std::io::{BufReader, Read};
use tea::io::Reader;

let f = File::open("foo.txt").ok().unwrap();
let mut decrypt_f = Reader::new(BufReader::new(f),
                                [1, 2, 3, 4], [5, 6]);
let mut s = "".to_string();
decrypt_f.read_to_string(&mut s).ok().unwrap();

Methods

impl<R: BufRead> Reader<R>

fn new(source: R, key: Key, iv: Block) -> Reader<R>

Wraps source in a Reader that will decrypt with the given key and iv (initialization vector).