Struct tea::io::Writer

pub struct Writer<W: Write> {
    // some fields omitted
}

Wraps an underlying std::io::Write so that bytes written get encrypted and passed through. You must call close() when finished writing to append the padding bytes.

Example:

use std::fs::File;
use std::io::Write;
use tea::io::Writer;

let f = File::create("foo.txt").ok().unwrap();
let mut crypt_f = Writer::new(f, [1, 2, 3, 4], [5, 6]);
crypt_f.write_all(b"Hello, world!").ok().unwrap();
crypt_f.close().ok().unwrap();

Methods

impl<W: Write> Writer<W>

fn new(sink: W, key: Key, iv: Block) -> Writer<W>

Wraps sink in a Writer that will encrypt with the given key and iv (initialization vector).

fn close(self) -> Result<W>

Writes the final padding bytes according to PKCS#7, destroys the encrypting wrapper, and returns the underlying std::io::Write object.