Struct tea::BlockCipher

pub struct BlockCipher {
    // some fields omitted
}

Adapts encipher and decipher into a block cipher using CBC.

Methods

impl BlockCipher

fn new(key: Key, iv: Block) -> BlockCipher

Creates an XTEA BlockCipher with a key and an initialization vector.

fn encrypt(&mut self, input: &mut BufRead, output: &mut Write)

Reads bytes from input until it's consumed, and writes the encrypted bytes to output. Will pad to an 8-byte boundary.

fn decrypt(&mut self, input: &mut BufRead, output: &mut Write)

Reads encrypted bytes from input until it's consumed, and writes the decrypted bytes to output. If the input needed to be padded, the result will have trailing zeroes.